Provided bug fixes making the code pass more tests

+++ db/migrate/20180220150734_remove_que_jobs.rb
@@ -1,5 +1,7 @@
 class RemoveQueJobs < ActiveRecord::Migration[5.1]
   def change
-    drop_table :que_jobs
+    if defined? Que
+      drop_table :que_jobs
+    end
   end
 end
